requirements.txt
Django==1.10.6
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
